Add further information on how to configure your deployment enviroment
Addressing #19, to improve how an architect/devops would configure their Kabanero
CRD to use a stack hub.

This section is the same as the deployment section in the "Working with stacks" guide, which
is also updated to extend the deploymeny section.

Other improvements in this PR for "Working with stacks":

- Change title of guide to "Customizing application stacks", which is more accurate.
- Add further information for how developers would customize their local dev environments with links to other guides
- Fix ordered lists by adjusting indenting.

Closes: #19

Signed-off-by: Sue Chaplain <sue_chaplain@uk.ibm.com>