Editorial review of guide
Copy of https://github.com/kabanero-io/guide-nexus-repository-integration/pull/9
from old repo.

General editing update: active voice, sentence simplification, etc.
In order to work with the new collective guides repo, the examples needed
to be zipped for download as it is no longer possible to clone the directory.

Moving back to /draft as not ready  for 0.9.0

Signed-off-by: Sue Chaplain <sue_chaplain@uk.ibm.com>